requirements.txt
# Core Framework
fastapi==0.123.9
uvicorn[standard]==0.38.0
starlette==0.50.0

# Database
motor==3.5.0
pymongo>=4.6,<5.0
dnspython==2.8.0

# Authentication & Security
passlib[bcrypt]==1.7.4
bcrypt==5.0.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
PyJWT==2.10.1

# Data Validation
pydantic==2.12.5
pydantic-core==2.41.5
email-validator==2.3.0
annotated-types==0.7.0

# Real-time and WebRTC signalling
python-socketio==5.15.0
python-engineio==4.12.3
simple-websocket==1.1.0

# File uploads / static files
python-multipart==0.0.20
aiofiles==25.1.0

# HTTP & Networking
httpx==0.28.1
httpcore==1.0.9
httptools==0.7.1

# WebSocket support
websockets==15.0.1
wsproto==1.3.2

# Configuration & Environment
python-dotenv==1.0.0

# Utilities & Dependencies
click==8.3.1
colorama==0.4.6
certifi==2025.11.12
idna==3.11
h11==0.16.0
anyio==4.12.0
bidict==0.23.1
exceptiongroup==1.3.1
typing-extensions==4.15.0
typing-inspection==0.4.2
watchfiles==1.1.1
setuptools==65.5.0
pip==22.3.1

# Cryptography
cffi==2.0.0
pycparser==2.23

# Utilities
PyYAML==6.0.3
annotated-doc==0.0.4

# Development / debugging helpers (optional)
# uvloop>=0.17.0  # Windows not supported - only use on Linux/macOS